Spring Boot Application for BrewDay!
Creating a spring boot based application for Brew Day

Contains JPA module that integrates with PostGreSQL for performing crud operations. The same is implemented for one resource - 'Recipe'. Uses Repository pattern and generates SQL statements on the fly
A Rest Controller for Recipe resource is also implemented. Finer details of the implementation will be implemented once the contract with UI is finalized
Contains security module to make sure the rest endpoints are served only Authenticated users are accessing. The exact approach is to be finalized with UI developer.

Please make sure you commit future enhancements into this container so that we are all working towards one application instead of capabilities spread across the branch which will make integration difficult.